Moved tests to Tests.Models folder
