+++ BankManager_Csharp/BankManager_Csharp/BankManager_Csharp.csproj
@@ -120,8 +120,9 @@
     <Compile Include="Models\Response.cs" />
     <Compile Include="Models\Transaction.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Tests\AccountTest.cs" />
-    <Compile Include="Tests\BankManagerTest.cs" />
+    <Compile Include="Tests\Controllers\LoginControllerTest.cs" />
+    <Compile Include="Tests\Models\AccountTest.cs" />
+    <Compile Include="Tests\Models\BankManagerTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="Views\web.config" />
@@ -143,6 +144,9 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
+  </ItemGroup>
+  <ItemGroup>
+    <Service Include="{82A7F48D-3B50-4B1E-B82E-3ADA8210C358}" />
   </ItemGroup>
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
